Refactoring of blending classes into a module and separating out the TriangularBlending across adjacent poitns
